% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QRRT_V6.R
\name{QRRT}
\alias{QRRT}
\title{QRRT}
\usage{
QRRT(Formula, Data, Disperse = 1, beta = NULL, n_times = 1, m = 8,
  n_orange = 25, n_green = c(6, 7, 4, 2, 2, 1, 1, 1, 1))
}
\arguments{
\item{Formula}{same format as linear model (lm)}

\item{Data}{need a data frame}

\item{Disperse}{standard deviation of the random starting value}

\item{beta}{starting value}

\item{n_times}{number of random starting values}

\item{m}{max number on green balls}

\item{n_orange}{number of orange balls}

\item{n_green}{number of green balls}
}
\value{
Data frame contains estimated beta, hypothesis result
}
\description{
Analysis QRRT data
}
\examples{
true_beta = c(log(5), 0, log(4), log(2), log(1))
n = 250 #sample size
age = 15:70
Age = sample(age, size = n, replace = TRUE)/100
continuous_x = rnorm(n, 0, 1)
discrete_x = sample(c("A", "B", "C"), size = n, replace = TRUE)
x = matrix(c(rep(1,n), Age, continuous_x, Age*continuous_x, as.factor(discrete_x)), ncol = 5)
Data = simulation_fun(n = 250, m = 8, n_orange = 25,n_green = c(6,7,4,2,2,1,1,1,1),
x = x, true_beta = true_beta)
res = QRRT(Formula = Y~.-1, Data = Data, Disperse = 1, beta = NULL, n_times = 1)

}
\author{
Meng Cao, F. Jay Breidt
}
